requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
ndg-httpsclient==0.5.1
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyOpenSSL==24.2.1
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.5
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
